src/main/java/com/hyeonah/urlshorteningservice/common/UrlValidator.java
package com.hyeonah.urlshorteningservice.common;

import java.util.regex.Matcher;
import java.util.regex.Pattern;

public class UrlValidator {
    private static final String URL_REGEX = "^(https?):\\/\\/([^:\\/\\s]+)(:([^\\/]*))?((\\/[^\\s/\\/]+)*)?\\/?([^#\\s\\?]*)(\\?([^#\\s]*))?(#(\\w*))?$";
    private static final Pattern URL_PATTERN = Pattern.compile(URL_REGEX);

    private UrlValidator() {
    }

    public static boolean validateUrl(final String url) {
        final Matcher m = URL_PATTERN.matcher(url);
        return m.matches();
    }
}
